Import initial version
Although the PCIe mode is described and the relevant drivers are built into
the kernel, the 5G modem only works in USB mode currently. In PCIe mode,
the /dev/wwan0mbim0 character device is created without a corresponding
wwan0 network device.

All other features are working as documented in README.