Update to linux 6.10.2 and u-boot 2024-07
Add ucom to package list for configuring modem and adjust for the rename
of the hostap package to hostapd. The openntpd package is now based on
openntpd-arachsys so use the ntpd -s feature in the router /etc/init.

Backport the correct fix for the gpio-hog issue, as contributed to
upstream u-boot. Update linux nvmem.diff to use the bdev_partno accessor
now direct access to bdev->bd_partno is not possible. Drop precal.diff
as it now included upstream.

Start an agetty on the serial console in the router /etc/init. It uses very
little memory and is handy for testing given the quiet log-level default.

The 5G modem still doesn't work correctly in PCIe mode. /dev/wwan0mbim0,
/dev/wwan0qcdm0 and /dev/wwan0qmi0 are all created and working but there
is no corresponding wwan0 network device.