Update to linux 6.10.3 and fix modem PCIe mode
The 5G modem was being incorrectly recognised under a general Qualcomm
vendor/product combination. Add a specific subvendor/subproduct rule for
the variant of the the RM520N-GL in the GL-X3000, which fixes the wwan
interface in PCIe mode.

Downgrade a harmless rate-limited warning to a debug message, matching
the handling of the same minor issue by the USB cdc-ncm driver.

Clarify in README that the init script and tinymbim work in either USB
or PCIe mode without modification.