Update to linux 6.14.4, uboot 2025.04 and mtk-atf 20250212
Adapt config/u-boot.dts to upstream u-boot changes. Reduce the maximum
eMMC speed to 26MHz and increase the SPI drive strength to 12mA following
GL-iNet's own fix for an issue with early hardware revisions.

Drop patches/alignment.diff as this is now in mainline tf-a and mtk-atf.

Drop patches/gpio.diff as this is now in mainline u-boot.

Rebase and refresh other patches.